[#96]Refactor(error): 명확한 타입 확인 위해 ErrorResponse 분리, ApiResponse에서 예외 시 명식적으로 반환하도록 함.
- Todo: ApiResponse ErrorResponse generic 수정 필요.
