+++ src/main/java/today/sesac/versebyverse/global/response/ApiResponse.java
@@ -1,6 +1,7 @@
 package today.sesac.versebyverse.global.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -21,12 +22,12 @@ public class ApiResponse<T> {
     private final Boolean isSuccess;
 
     /**
-     * API 요청의 성공 결과
+     * API 요청의 성공 결과.
      */
     private final T result;
 
     /**
-     * API 요청의 오류 내용
+     * API 요청의 오류 내용.
      */
     private final T error;
 
@@ -34,7 +35,7 @@ public class ApiResponse<T> {
      * API 요청 성공 시 성공 결과를 반환하기 위해 사용합니다.
      *
      * @param result API 요청의 성공 결과
-     * @return 성공 결과가 담긴 ApiResponse<T>
+     * @return 성공 결과가 담긴 {@code ApiResponse<T>}
      */
     public static <T> ApiResponse<T> success(T result) {
 
@@ -45,9 +46,20 @@ public class ApiResponse<T> {
      * API 요청 실패 시 오류를 알려주기 위해 사용합니다.
      *
      * @param errorResponse API 요청의 오류 내용
-     * @return 오류 내용이 담긴 ApiResponse<Void>
+     * @return 오류 내용이 담긴 {@code ApiResponse<ErrorResponse>}
      */
-    public static <T> ApiResponse<T> fail(T errorResponse) {
+    public static ApiResponse<ErrorResponse> fail(ErrorResponse errorResponse) {
+
+        return new ApiResponse<>(false, null, errorResponse);
+    }
+
+    /**
+     * API 요청 실패 시 오류를 알려주기 위해 사용합니다.
+     *
+     * @param errorResponse API 요청의 오류 내용
+     * @return 오류 내용이 담긴 {@code ApiResponse<List<ErrorResponse>>}
+     */
+    public static ApiResponse<List<ErrorResponse>> fail(List<ErrorResponse> errorResponse) {
 
         return new ApiResponse<>(false, null, errorResponse);
     }
